working on dijkstra

+++ lab3/src/algo.hpp
@@ -15,23 +15,23 @@ class Graph{
         /**
          * Temp debug constructor
          */
-        Graph(size_t V);
+        Graph(int V);
         
-        size_t V; /* Number of vertices */
-        size_t E; /* Nmber of edges */
-        size_t max_weight; /* Highest weight of all edges */
+        int V; /* Number of vertices */
+        int E; /* Nmber of edges */
+        int max_weight; /* Highest weight of all edges */
 
         /**
          * Adjacency list representing a graph.
          * Edges and weights are stored in form of a pair (vertex, weight)
          */ 
-        std::vector<std::vector<std::pair<size_t, size_t>>> adjacency_list;
+        std::vector<std::vector<std::pair<int, int>>> adjacency_list;
 
         /**
          * Ads an edge to the adjacency list
          * with a weight
          */
-        void add_edge(size_t a, size_t b, size_t weight);
+        void add_edge(int a, int b, int weight);
 
         /**
          * Neatly prints the adjacency list
@@ -41,25 +41,30 @@ class Graph{
         /**
          * Parses a text file to create a list of nodes
          * that will be sources in the ss variant of pahfinding
+         * @param ss_path path to the input file
          */
-        std::vector<size_t> ss_from_file(std::string ss_path);
+        std::vector<int> ss_from_file(std::string ss_path);
 
         /**
-         * Parses a text file to create a list of pairs of vertices
+         * @brief Parses a text file to create a list of pairs of vertices
          * between which we will find the shortest path
+         * @param p2p_path path to the input file
          */
-        std::vector<std::pair<size_t, size_t>> p2p_from_file(std::string p2p_path);
+        std::vector<std::pair<int, int>> p2p_from_file(std::string p2p_path);
 
         /**
-         * Finds shortest paths between source and all other nodes
-         * Assumes that:
-         *      - wieghts have integer, non-negative values
+         * @brief Finds shortest paths between source and all other nodes
+         * 
+         * Assumes that wieghts are integer, non-negative values
+         * @param source source node
+         * @return a vector of shortest path distances from source to other vertices
          */
-        std::vector<size_t> dijkstra(size_t source);
+        std::vector<int> dijkstra(int source);
 
-        std::vector<size_t> dial(size_t source);
+        std::vector<int> dial(int source);
 
-        std::vector<size_t> radixheap(size_t source);
+        std::vector<int> radixheap(int source);
+
 };
 
 #endif /* SRC_ALGO_INCLUDED */
